Move RecordServer into test
