+++ potboiler_common/src/test/mod.rs
@@ -1,3 +1,5 @@
 mod server_thread;
+mod record_server;
 
+pub use record_server::{RecordServer, RecordRequest};
 pub use server_thread::ServerThread;
